AVR fixups.

+++ libraries/UHS_host/UHS_host_INLINE.h
@@ -980,7 +980,7 @@ uint8_t UHS_USB_HOST_BASE::ctrlReq(uint8_t addr, uint64_t Request, uint16_t nbyt
 #if UHS_DEVICE_WINDOWS_USB_SPEC_VIOLATION_DESCRIPTOR_DEVICE
 
                                 // Should only be used for GET_DESCRIPTOR USB_DESCRIPTOR_DEVICE
-                                if(!addr && ((Request & 0xFF00FF00U) == ((USB_REQUEST_GET_DESCRIPTOR << 8) | (USB_DESCRIPTOR_DEVICE << 24)))) {
+                                if(!addr && ((Request & (uint32_t)0xFF00FF00U) == (((uint32_t)USB_REQUEST_GET_DESCRIPTOR << 8) | ((uint32_t)USB_DESCRIPTOR_DEVICE << 24)))) {
                                         HOST_DUBUG("ctrlReq3: acceptBuffer sz %i nbytes %i left %i\n\r", read, nbytes, left);
                                         left = 0;
                                         break;
